+++ hypervisor/arch/x86/trampoline.c
@@ -7,6 +7,7 @@
 #include <reloc.h>
 #include <trampoline.h>
 #include <vm0_boot.h>
+#include <e820.h>
 
 uint64_t trampoline_start16_paddr;
