+++ hypervisor/include/arch/x86/multiboot.h
@@ -80,4 +80,7 @@ struct multiboot_module {
 
 int parse_hv_cmdline(void);
 int init_vm_boot_info(struct acrn_vm *vm);
+
+/* boot_regs store the multiboot header address */
+extern uint32_t boot_regs[2];
 #endif
